cleanup getting descriptors/handles for ports.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@31288 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSMessagePort.m
@@ -1405,7 +1405,7 @@ typedef	struct {
  * This is a callback method used by the NSRunLoop class to determine which
  * descriptors to watch for the port.
  */
-- (void) getFds: (int*)fds count: (int*)count
+- (void) getFds: (NSInteger*)fds count: (NSInteger*)count
 {
   NSMapEnumerator	me;
   void			*sock;
